multi_agents/pyproject.toml
[project]
name = "multi-agents-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.21.0",
    "veadk-python==0.5.20",
    "veadk-python[extensions]==0.5.20",
    "agentkit-sdk-python>=0.5.1",
    "python-dotenv>=1.0.0",
]
